+++ docs/welcome/changelog.rst
@@ -9,6 +9,8 @@ Changelog
 Next
 ====
 
+* :bug:`119` - SampledDepletedMaterial now respects the value of `xUnits` 
+  - :squashed:`120`
 * :pull:`114` - Standalone branches in the coefficient files are stored
   and accessed using a single string, rather than a single-entry tuple
   ``branches['myBranch']`` vs. ``branches[('myBranch', )]``